CDAP-1152 Deprecate OrderedTable
