+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/dataset2/module/lib/hbase/HBaseTableModule.java
@@ -18,7 +18,6 @@ package co.cask.cdap.data2.dataset2.module.lib.hbase;
 
 import co.cask.cdap.api.dataset.module.DatasetDefinitionRegistry;
 import co.cask.cdap.api.dataset.module.DatasetModule;
-import co.cask.cdap.api.dataset.table.OrderedTable;
 import co.cask.cdap.api.dataset.table.Table;
 import co.cask.cdap.data2.dataset2.lib.table.hbase.HBaseTableDefinition;
 
@@ -28,11 +27,6 @@ import co.cask.cdap.data2.dataset2.lib.table.hbase.HBaseTableDefinition;
 public class HBaseTableModule implements DatasetModule {
   @Override
   public void register(DatasetDefinitionRegistry registry) {
-    // TODO: remove OrderedTable registration when OrderedTable is removed
-    registry.add(new HBaseTableDefinition("orderedTable"));
-    // so that it can be resolved via @Dataset
-    registry.add(new HBaseTableDefinition(OrderedTable.class.getName()));
-
     registry.add(new HBaseTableDefinition("table"));
     registry.add(new HBaseTableDefinition(Table.class.getName()));
   }
