+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/dataset2/module/lib/leveldb/LevelDBTableModule.java
@@ -18,7 +18,6 @@ package co.cask.cdap.data2.dataset2.module.lib.leveldb;
 
 import co.cask.cdap.api.dataset.module.DatasetDefinitionRegistry;
 import co.cask.cdap.api.dataset.module.DatasetModule;
-import co.cask.cdap.api.dataset.table.OrderedTable;
 import co.cask.cdap.api.dataset.table.Table;
 import co.cask.cdap.data2.dataset2.lib.table.leveldb.LevelDBTableDefinition;
 
@@ -28,11 +27,6 @@ import co.cask.cdap.data2.dataset2.lib.table.leveldb.LevelDBTableDefinition;
 public class LevelDBTableModule implements DatasetModule {
   @Override
   public void register(DatasetDefinitionRegistry registry) {
-    // TODO: remove OrderedTable registration when OrderedTable is removed
-    registry.add(new LevelDBTableDefinition("orderedTable"));
-    // so that it can be resolved via @Dataset
-    registry.add(new LevelDBTableDefinition(OrderedTable.class.getName()));
-
     registry.add(new LevelDBTableDefinition("table"));
     registry.add(new LevelDBTableDefinition(Table.class.getName()));
   }
